backend/requirements.txt
fastapi
uvicorn
pathway==0.29.0
python-jose
passlib[bcrypt]
psycopg2-binary
sqlalchemy
pydantic
ultralytics
openai
python-dotenv
python_multipart